+++ src/DependencyInjection/SonataNewsExtension.php
@@ -21,7 +21,6 @@ use Symfony\Component\DependencyInjection\Definition;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-use Twig\Extra\String\StringExtension;
 
 /**
  * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
@@ -104,7 +103,6 @@ class SonataNewsExtension extends Extension
 
         $this->configureClass($config, $container);
         $this->configureAdmin($config, $container);
-        $this->configureStringExtension($container);
     }
 
     /**
@@ -262,15 +260,5 @@ class SonataNewsExtension extends Extension
                 ],
             'orphanRemoval' => false,
         ]);
-    }
-
-    private function configureStringExtension(ContainerBuilder $container): void
-    {
-        if (!$container->hasDefinition('twig.extension.string') || !is_a($container->getDefinition('twig.extension.string')->getClass(), StringExtension::class)) {
-            $definition = new Definition(StringExtension::class);
-            $definition->addTag('twig.extension');
-
-            $container->setDefinition(StringExtension::class, $definition);
-        }
     }
 }
